test: ChipBadge 테스트 통합 및 접근성 보강
